# internal/adapters/storage/postgres/pggen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package pggen

import (
	"database/sql"
	"time"

	"github.com/google/uuid"
)

type LoyaltyCard struct {
	ID        uuid.UUID
	UserID    uuid.UUID
	Balance   sql.NullString
	CreatedAt time.Time
	UpdatedAt time.Time
	IsBlocked bool
}